Refactor the invite logic
